feat: use AppData for config and state in installed mode (#18)
* feat: sync gamepad camera target with OBS preview scene

When the OBS preview scene changes directly in OBS (not via X-Touch or
Stream Deck), automatically update the dynamic gamepad's camera target
to match. This prevents the scenario where the user thinks they're
controlling the camera visible in preview, but the gamepad is still
targeting the previous camera.

Changes:
- Add is_studio_mode() getter to ObsDriver
- Add broadcast_target_change() and public current_timestamp_millis() to API
- Add find_dynamic_gamepad_slot() helper in main.rs
- Add obs.currentPreviewScene signal handler in main.rs

Behavior:
- Only activates in studio mode (preview != program)
- Only affects cameras with enable_ptz: true
- Only updates if target is different (no redundant broadcasts)
- Notifies Stream Deck clients via WebSocket

Also adds code quality rules to CLAUDE.md.


* feat: use AppData for config and state in installed mode

Fix "Access Denied" errors when running from C:\Program Files by
detecting portable vs installed mode and using appropriate paths.

Changes:
- Add paths.rs module for path management
- Portable mode: requires explicit .portable marker file
- Installed mode (default): uses %APPDATA%\XTouch GW
- File logging gracefully falls back to console-only if directory
  creation fails
- Installer now places config.yaml in AppData instead of Program Files
- Exclude Stream Deck plugin logs from installer

Path layout in installed mode:
- Config: %APPDATA%\XTouch GW\config.yaml
- State:  %APPDATA%\XTouch GW\state\sled\
- Logs:   %APPDATA%\XTouch GW\logs\


* chore: code cleanup and version scripts

- Add version-bump.ps1 script for release management
- Add VERSION.txt for tracking current version
- Update make.ps1 with build improvements
- Code formatting and cleanup across modules
- Add dist/ to .gitignore


* fix: version consistency and paths.rs error handling

- Fix VERSION.txt to match Cargo.toml at 1.5.0
- Add anyhow::Context for better error messages on config copy failures
- Make diagnostic eprintln! calls debug-only to reduce release noise


* fix: include node_modules in Stream Deck plugin installer

The Stream Deck plugin crashed on startup in release mode because native
modules (canvas) were excluded from the installer. The canvas package
requires prebuilt binaries (.node files and DLLs) at runtime.

Changes:
- build-installer.ps1: After building, reinstall production-only deps
  to avoid bundling dev dependencies bloat
- xtouch-gw.iss: Remove node_modules from Excludes list

This increases installer size (~2MB -> ~27MB) but ensures the plugin
works when installed from the release package.


---------

+++ src/router/xtouch_input.rs
@@ -126,11 +126,8 @@ impl super::Router {
             let lsb = raw[1] & 0x7F;
             let msb = raw[2] & 0x7F;
             let value14 = ((msb as u16) << 7) | (lsb as u16);
-            debug!(
-                "← User moved fader: ch={} value14={}",
-                channel, value14
-            );
-            self.fader_setpoint.schedule(channel as u8, value14, None);
+            debug!("← User moved fader: ch={} value14={}", channel, value14);
+            self.fader_setpoint.schedule(channel, value14, None);
         }
 
         // Get active page and find control configuration
@@ -144,7 +141,8 @@ impl super::Router {
 
         // Check page-specific controls first, then global controls as fallback
         let config = self.config.read().await;
-        let control_config = page.controls
+        let control_config = page
+            .controls
             .as_ref()
             .and_then(|controls| controls.get(control_id))
             .or_else(|| {
@@ -170,12 +168,14 @@ impl super::Router {
 
         // Check if this is MIDI direct mode (send raw MIDI to bridge)
         if let Some(target_spec) = &control_config.midi {
-            self.handle_midi_direct_mode(raw, control_id, &control_config, target_spec).await;
+            self.handle_midi_direct_mode(raw, control_id, &control_config, target_spec)
+                .await;
             return;
         }
 
         // Driver action mode
-        self.handle_driver_action_mode(raw, control_id, &control_config).await;
+        self.handle_driver_action_mode(raw, control_id, &control_config)
+            .await;
     }
 
     /// Handle MIDI direct mode (transform and send to bridge)
@@ -208,8 +208,7 @@ impl super::Router {
         let target_msg = match target_spec.midi_type {
             crate::config::MidiType::Cc => {
                 if let (Some(ch), Some(cc)) = (target_spec.channel, target_spec.cc) {
-                    let value =
-                        crate::midi::convert::from_percent_7bit(normalized_value * 100.0);
+                    let value = crate::midi::convert::from_percent_7bit(normalized_value * 100.0);
                     Some(crate::midi::MidiMessage::ControlChange {
                         channel: ch.saturating_sub(1), // Config is 1-based, internal is 0-based
                         cc,
@@ -243,8 +242,7 @@ impl super::Router {
             },
             crate::config::MidiType::Pb => {
                 if let Some(ch) = target_spec.channel {
-                    let value =
-                        crate::midi::convert::from_percent_14bit(normalized_value * 100.0);
+                    let value = crate::midi::convert::from_percent_14bit(normalized_value * 100.0);
                     Some(crate::midi::MidiMessage::PitchBend {
                         channel: ch.saturating_sub(1),
                         value,
@@ -348,7 +346,10 @@ impl super::Router {
         if type_nibble == 0x9 && raw.len() >= 3 {
             let velocity = raw[2];
             if velocity == 0 {
-                debug!("Ignoring Note Off (velocity 0) for control '{}'", control_id);
+                debug!(
+                    "Ignoring Note Off (velocity 0) for control '{}'",
+                    control_id
+                );
                 return;
             }
         }
@@ -363,9 +364,10 @@ impl super::Router {
         // This allows drivers to receive a Number instead of raw bytes array
         if raw.len() >= 3 {
             let value = match type_nibble {
-                0x9 => raw[2] as f64,        // Note On: velocity
-                0xB => raw[2] as f64,        // Control Change: value
-                0xE => {                      // Pitch Bend: 14-bit value (0-16383)
+                0x9 => raw[2] as f64, // Note On: velocity
+                0xB => raw[2] as f64, // Control Change: value
+                0xE => {
+                    // Pitch Bend: 14-bit value (0-16383)
                     let lsb = raw[1] as u16;
                     let msb = raw[2] as u16;
                     ((msb << 7) | lsb) as f64
@@ -409,4 +411,3 @@ impl super::Router {
         }
     }
 }
-
